Separate asset browser view from window
